integration: extract test_unprepared_statement
Test `test_unprepared_statement()` is extracted into
`statements/unprepared.rs`.
